sentinel/scripts/requirements.txt
polygon-api-client>=1.13.0
redis>=5.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
ib_insync>=0.9.86
prometheus-client>=0.19.0
fredapi>=0.5.1
finnhub-python>=2.4.19
python-telegram-bot>=20.7
schedule>=1.2.0
